moved over edit owner profile and renamed to projectController
